Rename data to notifications
