+++ libs/desktop-shared/src/lib/index.ts
@@ -1,4 +1,4 @@
 export * from './api';
-export * from './data';
 export * from './logging';
 export * from './lokalisation';
+export * from './notifications';
